0.6.3

+++ game/game.h
@@ -26,7 +26,7 @@ typedef struct {
 typedef struct {
     player_tile_t tiles[PLAYER_TILES];  // tiles in player's hand
     char word[BOARD_SIZE+1];            // word which player tries to insert
-    short word_status[BOARD_SIZE];      // 0 if player doesn't have letter, 1 if has, -1 if it's blank
+    int word_status[BOARD_SIZE];      // 0 if player doesn't have letter, 1 if has, -1 if it's blank
                                         // also stores bonus multiplier for letter
     short word_orientaion = VERTICAL;   // vertical or horizontal
 } player_t;
